% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCProcessor.R
\name{compute_auc,AUCProcessor-method}
\alias{compute_auc,AUCProcessor-method}
\title{Compute activity scores for all cells}
\usage{
\S4method{compute_auc}{AUCProcessor}(
  object,
  sort_by = "expression",
  select_top_k = NULL,
  percent_of_target = 1,
  verbose = TRUE
)
}
\arguments{
\item{object}{A \code{\linkS4class{AUCProcessor}} object.}

\item{sort_by}{sorting criterion ("expression" or "weight")}

\item{select_top_k}{keep only top K targets per TF (NULL = use all)}

\item{percent_of_target}{percentage of targets to use (0-1)}

\item{verbose}{print progress messages}
}
\value{
A \code{\linkS4class{AUCProcessor}} object with computed scores
in the \code{auc_results} slot.
}
\description{
Efficiently calculates TF activity scores for each cell using its
corresponding label-specific GRN. Uses BiocParallel for parallelization.
}
