% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCProcessor.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute activity scores for an object}
\usage{
compute_auc(object, ...)
}
\arguments{
\item{object}{An object that implements a \code{compute_auc} method.}

\item{...}{Additional method-specific arguments.}
}
\value{
Object with computed activity scores.
}
\description{
S4 generic for computing TF activity scores (AUC-like metrics).
}
\examples{
weights_list <- list("0" = as.data.frame(matrix(rnorm(90), nrow = 3, 
                             dimnames = list(paste0("TF", 1:3), 
                                             paste0("Gene", 1:30)))),
                     "1" = as.data.frame(matrix(rnorm(90), nrow = 3, 
                             dimnames = list(paste0("TF", 1:3), 
                                             paste0("Gene", 1:30)))))
cell_labels <- data.frame(cell= c("Cell1", "Cell2"), label = c("0", "1"))
expr <- matrix(rnorm(60), nrow=2, byrow = FALSE, 
                dimnames = list(c("Cell1", "Cell2"),paste0("Gene", 1:30)))

processor <- AUCProcessor(weights     = weights_list,
                          expression  = expr,
                          cell_labels = cell_labels)
processor <- compute_auc(processor, sort_by = "expression")
}
