% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{.convert_weights_list_to_df}
\alias{.convert_weights_list_to_df}
\title{Convert a list-format weights of tf x target weights
to a long-format data.frame}
\usage{
.convert_weights_list_to_df(weights_list)
}
\arguments{
\item{weights_list}{A named list of data.frames (TFs in rows, targets in columns).
Each data.frame corresponds to a label/phenotype and should have rownames as TFs and colnames as targets.}
}
\value{
A data.frame with columns: tf, target, weight, and label (if multiple matrices in the list).
}
\description{
Convert a list-format weights of tf x target weights
to a long-format data.frame
}
\keyword{internal}
