% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_auc.R
\name{.subset_auc_for_label}
\alias{.subset_auc_for_label}
\title{Subset collected AUC for a specific label}
\usage{
.subset_auc_for_label(x, label)
}
\description{
Subset collected AUC for a specific label
}
\keyword{internal}
