% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCProcessor.R
\name{.validate_auc_processor_inputs}
\alias{.validate_auc_processor_inputs}
\title{Validate AUCProcessor inputs}
\usage{
.validate_auc_processor_inputs(weights, expression, cell_labels)
}
\arguments{
\item{weights}{list of weight matrices}

\item{expression}{expression matrix (genes x cells)}

\item{cell_labels}{cell-to-label mapping}
}
\description{
Validate AUCProcessor inputs
}
