% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar.R
\name{.weights_to_long}
\alias{.weights_to_long}
\title{Convert weight list to long-format data frame}
\usage{
.weights_to_long(x)
}
\arguments{
\item{x}{SimiCvizExperiment}
}
\value{
data.frame with columns: tf, target, weight, label, label_name
}
\description{
Convert weight list to long-format data frame
}
\keyword{internal}
