% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{export_network_pdf}
\alias{export_network_pdf}
\title{Export network plot to PDF}
\usage{
export_network_pdf(
  x,
  out_dir,
  prefix = "SimiCviz",
  condition = NULL,
  top_n = 500,
  width = 7,
  height = 7,
  overwrite = FALSE
)
}
\arguments{
\item{x}{\code{\link[=SimiCvizExperiment]{SimiCvizExperiment()}} or weights data.frame.}

\item{out_dir}{root output directory.}

\item{prefix}{filename prefix.}

\item{condition}{optional condition filter.}

\item{top_n}{integer; passed to \code{\link[=plot_network]{plot_network()}}.}

\item{width, height}{PDF dimensions (in inches).}

\item{overwrite}{logical; overwrite if file exists.}
}
\value{
Invisibly, the PDF filepath.
}
\description{
Export network plot to PDF
}
