% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCProcessor.R
\name{get_auc}
\alias{get_auc}
\title{Extract activity scores from an object}
\usage{
get_auc(object, ...)
}
\arguments{
\item{object}{An object that implements a \code{get_auc} method.}

\item{...}{Additional method-specific arguments.}
}
\value{
Extracted activity scores.
}
\description{
S4 generic for retrieving computed TF activity scores.
}
\examples{
#' # Generate minimal example data
weights_list <- list("0" = as.data.frame(matrix(rnorm(90), nrow = 3, 
                             dimnames = list(paste0("TF", 1:3), 
                                             paste0("Gene", 1:30)))),
                     "1" = as.data.frame(matrix(rnorm(90), nrow = 3, 
                             dimnames = list(paste0("TF", 1:3), 
                                             paste0("Gene", 1:30)))))
cell_labels <- data.frame(cell= c("Cell1", "Cell2"), label = c("0", "1"))
set.seed(123)
expr <- matrix(rnorm(60), nrow=2, byrow = FALSE, 
                dimnames = list(c("Cell1", "Cell2"),paste0("Gene", 1:30)))

processor <- AUCProcessor(weights     = weights_list,
                          expression  = expr,
                          cell_labels = cell_labels)
processor <- compute_auc(processor, sort_by = "expression")
auc_scores <- get_auc(processor)
head(auc_scores)
 #          TF1       TF2       TF3
 # Cell1 0.5308869 0.5307807 0.5618426
 # Cell2 0.5752393 0.5399460 0.5304443
}
