% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_expression_matrix}
\alias{load_expression_matrix}
\title{Load expression matrix from various sources /file formats
Allowed formats:
\itemize{
\item CSV file
\item h5ad file
\item rds file (Seurat or SingleCellExperiment object)
\item matrix or data.frame in R
}}
\usage{
load_expression_matrix(expression)
}
\arguments{
\item{expression}{file path, matrix, or data.frame}
}
\value{
expression matrix
}
\description{
Load expression matrix from various sources /file formats
Allowed formats:
\itemize{
\item CSV file
\item h5ad file
\item rds file (Seurat or SingleCellExperiment object)
\item matrix or data.frame in R
}
}
\examples{
expression_mat_path <- system.file("extdata", 
                       file.path("inputFiles", "example_expression.pickle"),
                       package = "SimiCviz")
expression_mat <- load_expression_matrix(expression_mat_path)
}
