% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_auc_csv}
\alias{read_auc_csv}
\title{Read SimiC-style AUC results from CSV (compatibility)}
\usage{
read_auc_csv(file, ...)
}
\arguments{
\item{file}{path to a CSV file containing AUC metrics.
Expected columns include at least \code{cell}, \code{tf}, and \code{score}.}

\item{...}{additional arguments passed to \code{\link[utils:read.table]{utils::read.csv()}}.}
}
\value{
data.frame with AUC metrics.
}
\description{
Read SimiC-style AUC results from CSV (compatibility)
}
