% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_auc_pickle}
\alias{read_auc_pickle}
\title{Read SimiCPipeline output AUC matrices (pickle file)}
\usage{
read_auc_pickle(file)
}
\arguments{
\item{file}{path to a pickle file containing per-label AUC matrices.}
}
\value{
A named list of data.frames / matrices (one per label).
}
\description{
Returns a list of matrices (one per label), each with cells in rows and
TFs in columns.
}
\examples{
# This file does not exist in the package due to size constrains, but if you have a SimiCPipeline output, you can use:
\donttest{
 # auc_file <- system.file("extdata",paste0("outputSimic/example_simic_wAUC.pickle"))
 # auc_list <- read_auc_pickle(auc_file)
}
}
