% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_pickle}
\alias{read_pickle}
\title{Read pickle files}
\usage{
read_pickle(file)
}
\arguments{
\item{file}{path to the pickle file}
}
\value{
object containing the contents of the pickle file
}
\description{
Read pickle files
}
\examples{
weights_file <- system.file("extdata",  
                file.path("outputSimic/example_simic_weights.pickle"), 
                package = "SimiCviz")
simic_weights <- read_pickle(weights_file)
}
