% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimiCviz-package.R
\name{setCellLabels}
\alias{setCellLabels}
\title{Set or update cell-to-label mapping}
\usage{
setCellLabels(x, cell_labels)
}
\arguments{
\item{x}{\code{SimiCvizExperiment}.}

\item{cell_labels}{Any format accepted by \code{\link{load_cell_labels}}.}
}
\value{
Modified \code{SimiCvizExperiment} object.
}
\description{
Set or update cell-to-label mapping
}
\examples{
 wt <- data.frame(tf     = c("TF1", "TF2"),
                   target = c("Gene1", "Gene2"), 
                   weight = c(0.5, 0.8),
                   label  = c(0, 1)
                   )
 simic <- SimiCvizExperiment(weights = wt)

 cell_labels <- data.frame(cell = c("Cell1", "Cell2"), 
                          label = c(0, 1))
 simic <- setCellLabels(simic, cell_labels)
}
