% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCProcessor.R
\name{AUCProcessor}
\alias{AUCProcessor}
\title{Initialize AUCProcessor}
\usage{
AUCProcessor(
  weights,
  expression,
  cell_labels,
  qc_type = NULL,
  qc_threshold = 0.7,
  adj_r2_list = NULL,
  n_cores = 1,
  backend = "sequential"
)
}
\arguments{
\item{weights}{list with weight matrices per label (if simic input)
or data.frame with columns:
tf, target, weight, label, tf (if data.frame input).
adj_p_val (optional, for p-value filtering 'qc_type' = "p_value").}

\item{expression}{expression matrix (cells x cells) or path to file.
\strong{IMPORTANT}: Expression must be in genes x cells format
(genes as rows, cells as columns), matching
Seurat/SingleCellExperiment conventions.}

\item{cell_labels}{data.frame with cell-to-label mapping}

\item{qc_type}{character type of quality metric.
Values: "adj_r2" or other column name (pvalue, adj_p_val, etc.)
provided in weights in data.frame format. If "adj_r2",
targets with R2 below threshold are filtered out.
If other (i.e. "p_value"), targets with qc_metric above threshold
are filtered out.}

\item{qc_threshold}{threshold for filtering targets based on
qc_metric (default: 0.7 for R2).}

\item{adj_r2_list}{(optional) A list of R2 values per label per
Only used if weights are provided in list format
qc_type is "adj_r2".}

\item{n_cores}{number of cores for parallelization (default: 1)}

\item{backend}{parallelization backend ("sequential", "multicore",
"multisession")}
}
\value{
A \code{\linkS4class{AUCProcessor}} object with initialized slots.
}
\description{
Initialize AUCProcessor
}
\examples{
weight_path <- system.file("extdata", "example_weights.csv",  
               package = "SimiCviz")
weights_df <- read_weights_csv(weight_path)
expression_mat_path <- system.file("extdata", 
                       file.path("inputFiles", "example1_expression.pickle"),
                       package = "SimiCviz")
cell_labels_path <- system.file("extdata", 
                     file.path("inputFiles", "treatment_annotation.csv"), 
                      package = "SimiCviz")
cell_labels <- load_cell_labels(cell_labels_path, header = TRUE, sep = ",")

AS_processor <- AUCProcessor(weights = weights_df,
                             expression = expression_mat_path,
                             cell_labels = cell_labels,
                             qc_type = "adj_p_val",
                             qc_threshold = 0.05,  # Filter targets above
                             n_cores = 2,
                             backend = "multisession")

}
