% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCProcessor.R
\name{compute_auc}
\alias{compute_auc}
\title{Compute activity scores for an object}
\usage{
compute_auc(object, ...)
}
\arguments{
\item{object}{An object that implements a \code{compute_auc} method.}

\item{...}{Additional method-specific arguments.}
}
\value{
Object with computed activity scores.
}
\description{
S4 generic for computing TF activity scores (AUC-like metrics).
}
