% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCProcessor.R
\name{get_auc}
\alias{get_auc}
\title{Extract activity scores from an object}
\usage{
get_auc(object, ...)
}
\arguments{
\item{object}{An object that implements a \code{get_auc} method.}

\item{...}{Additional method-specific arguments.}
}
\value{
Extracted activity scores.
}
\description{
S4 generic for retrieving computed TF activity scores.
}
