% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimiCviz-package.R
\name{is.SimiCvizExperiment}
\alias{is.SimiCvizExperiment}
\title{Check if object is a SimiCvizExperiment}
\usage{
is.SimiCvizExperiment(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical, TRUE if x is a SimiCvizExperiment.
}
\description{
Tests whether an object is of class SimiCvizExperiment.
}
