% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_weights_csv}
\alias{read_weights_csv}
\title{Read Long-style TF-target weights from CSV (compatibility)}
\usage{
read_weights_csv(file, ...)
}
\arguments{
\item{file}{path to a CSV file containing TF-target weights.
Expected minimal columns: \code{tf}, \code{target}, \code{weight}.
Additional columns such as \code{label}, etc. are preserved.}

\item{...}{additional arguments passed to \code{\link[utils:read.table]{utils::read.csv()}}.}
}
\value{
data.frame with weights.
}
\description{
Read Long-style TF-target weights from CSV (compatibility)
}
\examples{
weight_path <- system.file("extdata", "example_weights.csv",  package = "SimiCviz")
weights_df <- read_weights_csv(weight_path)
head(weights_df)
#       tf target    weight label
# 1   Pms1 Brinp3 0.0000000     0
# 2  Prrx2 Brinp3 1.5622762     0
# 3   Ets1 Brinp3 0.0000000     0
# 4   Tet2 Brinp3 0.8540945     0
# 5 Crebrf Brinp3 0.0000000     0
# 6   Gli3 Brinp3 0.0000000     0

}
