% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_weights_pickle}
\alias{read_weights_pickle}
\title{Read SimiCPipeline output weights (pickle file)}
\usage{
read_weights_pickle(file)
}
\arguments{
\item{file}{path to the pickle file containing SimiC weights.}
}
\value{
A list of length n_labels (phenotypes) with GRN weight matrices (TFs x targets).
}
\description{
Read SimiCPipeline output weights (pickle file)
}
