% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimiCviz-package.R
\name{setCellLabels}
\alias{setCellLabels}
\title{Set or update cell-to-label mapping}
\usage{
setCellLabels(x, cell_labels)
}
\arguments{
\item{x}{\code{SimiCvizExperiment}.}

\item{cell_labels}{Any format accepted by \code{\link{load_cell_labels}}.}
}
\value{
Modified \code{SimiCvizExperiment} object.
}
\description{
Set or update cell-to-label mapping
}
